requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
mysqlclient>=1.4.2
gunicorn
django-bootstrap-datepicker-plus>=3.0.5